docs(sample-02): upgrade sample to cordless v3
