+++ sample/02-basic-javascript/commands/ping.js
@@ -1,0 +1,7 @@
+const ping = {
+  name: 'ping',
+  description: 'Responds to your ping with a pong!',
+  handler: ({ interaction }) => interaction.reply('Pong!'),
+}
+
+export default ping
