--- sample/02-basic-javascript/functions/ping.js
@@ -1,6 +1,0 @@
-const ping = {
-  condition: (msg) => msg.content === 'ping',
-  callback: (msg) => msg.reply('pong'),
-}
-
-export default ping
