+++ sample/02-basic-javascript/package.json
@@ -13,11 +13,11 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "dependencies": {
-    "cordless": "^2.2.0",
+    "cordless": "^3.0.0",
     "dotenv": "^16.0.1"
   },
   "devDependencies": {
-    "jest": "^28.1.2"
+    "jest": "^28.1.3"
   },
   "jest": {
     "testEnvironment": "jest-environment-node",
